// src/models/aes.ts
export interface ProtectedDataAES {
    ciphertext: string;
    tag: string;
    iv: string;
}